+++ ScriptableRenderPipeline/HDRenderPipeline/HDRP/HDRenderPipeline.cs
@@ -1,15 +1,10 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine.Rendering;
 using UnityEngine.Experimental.Rendering;
 using System;
 using System.Diagnostics;
 using System.Linq;
 using UnityEngine.Rendering.PostProcessing;
-#if UNITY_EDITOR
-using UnityEditor;
-using UnityEditor.Experimental.Rendering;
-#endif
 
 namespace UnityEngine.Experimental.Rendering.HDPipeline
 {
@@ -145,47 +140,6 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 			CoreUtils.ResizeHTile(ref m_HTile, ref m_HTileRT, camera.renderTextureDesc);
         }
     }
-
-#if UNITY_EDITOR
-    public class SceneViewDrawMode
-    {
-        static private bool RejectDrawMode(SceneView.CameraMode cameraMode)
-        {
-            if (cameraMode.drawMode == DrawCameraMode.TexturedWire ||
-                cameraMode.drawMode == DrawCameraMode.ShadowCascades ||
-                cameraMode.drawMode == DrawCameraMode.RenderPaths ||
-                cameraMode.drawMode == DrawCameraMode.AlphaChannel ||
-                cameraMode.drawMode == DrawCameraMode.Overdraw ||
-                cameraMode.drawMode == DrawCameraMode.Mipmaps ||
-                cameraMode.drawMode == DrawCameraMode.DeferredDiffuse ||
-                cameraMode.drawMode == DrawCameraMode.DeferredSpecular ||
-                cameraMode.drawMode == DrawCameraMode.DeferredSmoothness ||
-                cameraMode.drawMode == DrawCameraMode.DeferredNormal ||
-                cameraMode.drawMode == DrawCameraMode.ValidateAlbedo ||
-                cameraMode.drawMode == DrawCameraMode.ValidateMetalSpecular ||
-                cameraMode.drawMode == DrawCameraMode.ShadowMasks ||
-                cameraMode.drawMode == DrawCameraMode.LightOverlap
-                )
-                return false;
-
-            return true;
-        }
-
-        static public void SetupDrawMode()
-        {
-            ArrayList sceneViewArray = SceneView.sceneViews;
-            foreach (SceneView sceneView in sceneViewArray)
-                sceneView.onValidateCameraMode += RejectDrawMode;
-        }
-
-        static public void ResetDrawMode()
-        {
-            ArrayList sceneViewArray = SceneView.sceneViews;
-            foreach (SceneView sceneView in sceneViewArray)
-                sceneView.onValidateCameraMode -= RejectDrawMode;
-        }
-    }
-#endif
 
     public class HDRenderPipeline : RenderPipeline
     {
