+++ ScriptableRenderPipeline/HDRenderPipeline/HDRP/SceneViewDrawMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ce13303c7796bd4d80d623d982e7460
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
